+++ Components/Hlms/Unlit/include/OgreHlmsUnlit.h
@@ -114,9 +114,9 @@ namespace Ogre
                                            uint32 lastCacheHash, CommandBuffer *commandBuffer,
                                            bool isV1 );
 
-        HlmsUnlit( Archive *dataFolder, ArchiveVec *libraryFolders, size_t constBufferSize );
+        HlmsUnlit( Archive *dataFolder, ArchiveVec *libraryFolders, uint32 constBufferSize );
         HlmsUnlit( Archive *dataFolder, ArchiveVec *libraryFolders, HlmsTypes type,
-                   const String &typeName, size_t constBufferSize );
+                   const String &typeName, uint32 constBufferSize );
 
     public:
         HlmsUnlit( Archive *dataFolder, ArchiveVec *libraryFolders );
