+++ OgreMain/include/Compositor/OgreCompositorShadowNode.h
@@ -220,7 +220,7 @@ namespace Ogre
         /// is no light that could be linked with that shadow map index.
         /// i.e. if isShadowMapIdxActive( shadowMapIdx ) is true, then we'll
         /// return a valid pointer.
-        const Light *getLightAssociatedWith( uint32 shadowMapIdx ) const;
+        const Light *getLightAssociatedWith( size_t shadowMapIdx ) const;
 
         /** Outputs the min & max depth range for the given camera. 0 & 100000 if camera not found
         @remarks
